first-ai-finance-assistance/requirements.txt
fastapi
uvicorn
openai>=1.10.0
sentence-transformers
faiss-cpu
numpy
python-dotenv
